+++ asdf/_tests/test_schema.py
@@ -1020,13 +1020,6 @@ def test_custom_validation_with_external_ref_bad(tmp_path):
         pass
 
 
-def test_load_schema_resolve_local_refs_deprecated():
-    custom_schema_path = helpers.get_test_data_path("custom_schema_definitions.yaml")
-
-    with pytest.deprecated_call():
-        schema.load_schema(custom_schema_path, resolve_local_refs=True)
-
-
 def test_nonexistent_tag(tmp_path):
     """
     This tests the case where a node is tagged with a type that apparently