+++ asdf/schema.py
@@ -16,7 +16,7 @@ from asdf._jsonschema.exceptions import RefResolutionError, ValidationError
 
 from . import constants, generic_io, reference, tagged, treeutil, util, versioning, yamlutil
 from .config import get_config
-from .exceptions import AsdfDeprecationWarning, AsdfWarning
+from .exceptions import AsdfWarning
 from .extension import _legacy
 from .util import patched_urllib_parse
 
@@ -230,7 +230,7 @@ class _ValidationContext:
 
 @lru_cache
 def _create_validator(validators=YAML_VALIDATORS, visit_repeat_nodes=False):
-    meta_schema = _load_schema_cached(YAML_SCHEMA_METASCHEMA_ID, _legacy.get_default_resolver(), False, False)
+    meta_schema = _load_schema_cached(YAML_SCHEMA_METASCHEMA_ID, _legacy.get_default_resolver(), False)
 
     type_checker = mvalidators.Draft4Validator.TYPE_CHECKER.redefine_many(
         {
@@ -389,7 +389,7 @@ def _make_resolver(url_mapping):
     )
 
 
-def load_schema(url, resolver=None, resolve_references=False, resolve_local_refs=False):
+def load_schema(url, resolver=None, resolve_references=False):
     """
     Load a schema from the given URL.
 
@@ -407,16 +407,7 @@ def load_schema(url, resolver=None, resolve_references=False, resolve_local_refs
     resolve_references : bool, optional
         If ``True``, resolve all ``$ref`` references.
 
-    resolve_local_refs : bool, optional
-        If ``True``, resolve all ``$ref`` references that refer to other objects
-        within the same schema. This will automatically be handled when passing
-        ``resolve_references=True``, but it may be desirable in some cases to
-        control local reference resolution separately.
-        This parameter is deprecated.
     """
-    if resolve_local_refs is True:
-        warnings.warn("The 'resolve_local_refs' parameter is deprecated.", AsdfDeprecationWarning)
-
     if resolver is None:
         # We can't just set this as the default in load_schema's definition
         # because invoking get_default_resolver at import time leads to a circular import.
@@ -425,7 +416,7 @@ def load_schema(url, resolver=None, resolve_references=False, resolve_local_refs
     # We want to cache the work that went into constructing the schema, but returning
     # the same object is treacherous, because users who mutate the result will not
     # expect that they're changing the schema everywhere.
-    return copy.deepcopy(_load_schema_cached(url, resolver, resolve_references, resolve_local_refs))
+    return copy.deepcopy(_load_schema_cached(url, resolver, resolve_references))
 
 
 def _safe_resolve(resolver, json_id, uri):
@@ -467,11 +458,11 @@ def _safe_resolve(resolver, json_id, uri):
 
 
 @lru_cache
-def _load_schema_cached(url, resolver, resolve_references, resolve_local_refs):
+def _load_schema_cached(url, resolver, resolve_references):
     loader = _make_schema_loader(resolver)
     schema, url = loader(url)
 
-    if resolve_references or resolve_local_refs:
+    if resolve_references:
 
         def resolve_refs(node, json_id):
             if json_id is None:
@@ -483,12 +474,8 @@ def _load_schema_cached(url, resolver, resolve_references, resolve_local_refs):
                 if suburl_base == url or suburl_base == schema.get("id"):
                     # This is a local ref, which we'll resolve in both cases.
                     subschema = schema
-                elif resolve_references:
-                    # Only resolve non-local refs when the flag is set.
-                    subschema = load_schema(suburl_base, resolver, True)
                 else:
-                    # Otherwise return the $ref unmodified.
-                    return node
+                    subschema = load_schema(suburl_base, resolver, True)
 
                 return reference.resolve_fragment(subschema, suburl_fragment)
 
@@ -742,7 +729,7 @@ def check_schema(schema, validate_default=True):
         applicable_validators = methodcaller("items")
 
     meta_schema_id = schema.get("$schema", YAML_SCHEMA_METASCHEMA_ID)
-    meta_schema = _load_schema_cached(meta_schema_id, _legacy.get_default_resolver(), False, False)
+    meta_schema = _load_schema_cached(meta_schema_id, _legacy.get_default_resolver(), False)
 
     resolver = _make_resolver(_legacy.get_default_resolver())